begin work on RingTx
